re-solve 'products not loaded'
When viewing product details without first navigating through all products, the
products would not be loaded, and the details page would crash. This was
previously solved by using a common parent component responsible for loading
the products, and using a render prop to render something depending on the
products. This problem was now re-solved by always loading the single product,
whose details were to be viewed.